automations/invoice-webapp/backend/requirements.txt
# Invoice Upload Web App - Backend (Lesson 2.4)
flask>=3.0.0
flask-cors>=4.0.0
pdfplumber>=0.10.0
requests>=2.28.0
python-dotenv>=1.0.0
